Moved queryOperators to a dedicated file
